Convert group label display to a custom fork of Mui DataGrid.
I needed a custom fork because DataGrid includes a RegExp with
Unicode properties, which are an ECMAScript 2018 feature. I changed
that one line in the fork, which is only used in editable DataGrids.

DataGrid gives us a lot more functionality than I would be able
to put together on my own.