Make upload errors more evident & improve logging.
Catch Exception rather than Throwable, because in general we
don't want to catch unrecoverable Errors without a good reason.

Change log/warns to log/error.

Removed the code that forces single-threaded behavior when *conn* is
set. `future` has "binding coveyance" which means that the futures
will see the same *conn* as the scope that creates them.