Use Input types for all mutations and subscriptions
We're making this breaking change now since we have only a couple of
dependents and we control them.

Having multiple arguments gets unwieldy very quickly. It's much
cleaner to have a single input argument. Even if an operation has only
one or two args currently, it may have more later and it's not
possible to change it to an input arg without breaking things. Also,
this is very consistent: Always pass an input arg.