Rename old ctgov external_ids
sysrev.datasource.core/save-article-data won't update article-data if
we have an existing one with the same external_id. That breaks the new
ctgov system because it needs different data there, and there are
already articles imported. To work around this, we rename all of the
existing ids. These shouldn't be needed for anything and could
possibly be deleted, but this way if it turns out that someone needs
them, we can migrate their project.