Don't write as often to project-member.access-time
This adds a considerable number of writes, particularly in
with-authorize. Since we're only using this to order the user's list
of projects, it is enough to record access-time when project-info is
accessed.

If we need more fine-grained access data, we should write it to
another datastore (like DynamoDB) to avoid load in postgres.