Add :sr-context component to simplify dependency injection
We have been using the :web-server component to find other components
like the config and postgres. This isn't ideal because many fns can be
called from non-web-server executors such as the scheduler. Also, few
of the fns taking web-server as an argument actually do anything with
the web-server. They use it merely as a way to look up the components
that they actually do care about.

We clean this up with a "sysrev context" (:sr-context) that acts as
a simple map of DI components. This decouples fns from their executor,
and provides a clear place to add future common components (like
caches and AWS service clients).