Use donut.system for sysrev's system map
donut.system has a much better design that allowing for improved
modularity.

This just wraps the existing com.stuartsierra components.
The only immediate change is that we don't have to be careful to
avoid the nil values that broke reviews today. donut.system just
treats it as a constant nil rather than throwing.