Factor out job queue logic
Move job queue logic to job-queue and job-runner components. This
changes it from tightly coupled code that depends on everything to
components that only depend on postgres and a passed-in function.

It also removes the need to manually run the job queue in API endpoints
by automatically kicking off queue processing when a job is added. The
polling job runner still works basically the same way as before and will
continue to handle retries.